backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
pypdf
python-dotenv==1.0.0
opencv-python==4.8.1.78
Werkzeug==3.0.1
numpy<2
Flask-Mail==0.9.1
google-genai
